requirements.txt
fastapi
uvicorn[standard]
python-socketio
python-dotenv
faker
google-generativeai
pydantic